refactor(sanity): move version provider to its own file + organise + add useVersion
