+++ packages/sanity/src/core/studio/StudioProvider.tsx
@@ -8,13 +8,13 @@ import jsx from 'refractor/lang/jsx.js'
 import typescript from 'refractor/lang/typescript.js'
 import {AddonDatasetProvider} from 'sanity'
 
-import {VersionProvider} from '../../_singletons/core/form/VersionContext'
 import {LoadingBlock} from '../components/loadingBlock'
 import {ErrorLogger} from '../error/ErrorLogger'
 import {errorReporter} from '../error/errorReporter'
 import {LocaleProvider} from '../i18n'
 import {ResourceCacheProvider} from '../store'
 import {UserColorManagerProvider} from '../user-color'
+import {VersionProvider} from '../versions/context/VersionProvider'
 import {ActiveWorkspaceMatcher} from './activeWorkspaceMatcher'
 import {AuthBoundary} from './AuthBoundary'
 import {ColorSchemeProvider} from './colorScheme'
