+++ packages/sanity/src/core/versions/components/BundleMenu.tsx
@@ -1,14 +1,11 @@
 import {CheckmarkIcon} from '@sanity/icons'
 import {Box, Flex, Menu, MenuButton, MenuDivider, MenuItem, Spinner, Text} from '@sanity/ui'
-import {type ReactElement, useCallback, useContext} from 'react'
+import {type ReactElement, useCallback} from 'react'
 import {RelativeTime} from 'sanity'
 import {styled} from 'styled-components'
 
-import {
-  VersionContext,
-  type VersionContextValue,
-} from '../../../_singletons/core/form/VersionContext'
 import {type BundleDocument} from '../../store/bundles/types'
+import {useVersion} from '../context/useVersion'
 import {LATEST} from '../util/const'
 import {isDraftOrPublished} from '../util/dummyGetters'
 import {BundleBadge} from './BundleBadge'
@@ -34,8 +31,7 @@ export function BundleMenu(props: BundleListProps): JSX.Element {
 
   // eslint-disable-next-line no-warning-comments
   // TODO MAKE SURE THIS IS HOW WE WANT TO DO THIS
-  const {currentVersion, setCurrentVersion, isDraft} =
-    useContext<VersionContextValue>(VersionContext)
+  const {currentVersion, setCurrentVersion, isDraft} = useVersion()
 
   // eslint-disable-next-line no-warning-comments
   // FIXME REPLACE WHEN WE HAVE REAL DATA
